vector-store: refactor CDC manager to `db_cdc` actor
This change moves the code to logically separate CDC manager
module which would allow to create and manage many CDC readers
with different configuration parameters (e.g. safety and sleep
intervals).
